// src/frontend/customer/src/api/templates.ts
export interface TemplateInfo {
  id: number
  authorId: number
  name: string
  description: string
  templateType: 'server_template' | 'world_gen' | 'modpack'
  category: string
  gameMode: string
  maxPlayers: number
  tags: string
  coverImage: string
  configJson: string
  modList: string
  version: number
  downloadCount: number
  ratingAvg: number
  ratingCount: number
  status: string
  verified: number
  createdAt: string
  updatedAt: string
}